+++ include/oc_rep.h
@@ -28,7 +28,7 @@ extern CborEncoder g_encoder, root_map, links_array;
 extern int g_err;
 
 void oc_rep_new(uint8_t *payload, int size);
-int oc_rep_finalize(void);
+ssize_t oc_rep_finalize(void);
 
 #define oc_rep_object(name) &name##_map
 #define oc_rep_array(name) &name##_array
